+++ delcache.sh
@@ -1,0 +1,19 @@
+if [ -d dialogExecution/__pycache__ ]; then
+    rm -r dialogExecution/__pycache__
+fi
+
+if [ -d errors/__pycache__ ]; then
+    rm -r errors/__pycache__
+fi
+
+if [ -d platformSpecific/__pycache__ ]; then
+    rm -r platformSpecific/__pycache__
+fi
+
+if [ -d translations/__pycache__ ]; then
+    rm -r translations/__pycache__
+fi
+
+if [ -d uiScripts/__pycache__ ]; then
+    rm -r uiScripts/__pycache__
+fi