+++ CHANGELOG.md
@@ -30,6 +30,7 @@ You now need the latest version of your Linux distribution or at least Windows 1
 - Thanks to Calibrato, we can now add the Polish translation.
 - You can now start the VM without passing a specific time and date to it.
 - The building process for Windows binaries with cx_Freeze is now documented in the readme.
+- Now you can delete the cache with the delcache scripts for your OS.
 
 ## Changes made compared to 1.2.0.5508_rc1